Fixing syslog formatting issues with were causing buffer overruns on 32-bit architectures (#10)
noddos startup script adds multicast route to eth0

+++ src/DnsmasqLogFile.h
@@ -57,8 +57,8 @@ struct DnsQuery {
  */
 class DnsmasqLogFile : public iDeviceInfoSource, public iCache {
 private:
-	std::map<uint64_t, std::shared_ptr<DnsLogEntry>> DnsQueryMap;
-    std::map<uint64_t, std::shared_ptr<DhcpRequest>> DhcpRequestMap;
+	std::map<unsigned long long, std::shared_ptr<DnsLogEntry>> DnsQueryMap;
+    std::map<unsigned long long, std::shared_ptr<DhcpRequest>> DhcpRequestMap;
 
     FILE *fp = nullptr;
     HostCache &hCache;
